+++ ConfigCommands-core/src/main/java/me/willkroboth/ConfigCommands/ConfigCommandsHandler.java
@@ -4,9 +4,9 @@ import dev.jorel.commandapi.CommandTree;
 import dev.jorel.commandapi.arguments.*;
 import dev.jorel.commandapi.executors.CommandExecutor;
 import dev.jorel.commandapi.executors.ExecutorType;
-import me.willkroboth.ConfigCommands.Commands.BuildCommandHandler;
-import me.willkroboth.ConfigCommands.Commands.FunctionCommandHandler;
-import me.willkroboth.ConfigCommands.Commands.ReloadCommandHandler;
+import me.willkroboth.ConfigCommands.SystemCommands.BuildCommandHandler;
+import me.willkroboth.ConfigCommands.SystemCommands.FunctionCommandHandler;
+import me.willkroboth.ConfigCommands.SystemCommands.ReloadCommandHandler;
 import me.willkroboth.ConfigCommands.Exceptions.RegistrationException;
 import me.willkroboth.ConfigCommands.HelperClasses.ConfigCommandAddOn;
 import me.willkroboth.ConfigCommands.HelperClasses.ConfigCommandBuilder;
