+++ ConfigCommands-core/src/main/java/me/willkroboth/ConfigCommands/SystemCommands/CommandContext.java
@@ -1,4 +1,4 @@
-package me.willkroboth.ConfigCommands.Commands;
+package me.willkroboth.ConfigCommands.SystemCommands;
 
 import org.bukkit.command.CommandSender;
